Move decorator-related logic into its own file
